Day 20
Took a very long time to understand what I was doing. Tried to make a NX graph where the cheat was encoded in the nodes. This worked for p1, but I couldn't figure out how to make it work for p2.

Wound up with a clean solution that checks all cheats for all nodes. If the cheat results in a savings of 100 or more picoseconds, it's counted.

In retrospect, I think I could have done it with simply edges on the NX graph. I'd just set the weight to the manhattan distance between the points and use the cost. Or I could have calculated distance myself instead of relying on `len(path)`. Ahh well. Maybe next time.